# docs/swagger-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lumina AI API",
    "description": "Lumina AI 图片处理应用后端API文档\n\n本API提供以下核心功能：\n- 用户认证与授权\n- AI问答与推荐\n- 图片上传与处理（单图/批量）\n- 作品管理\n- 订阅与支付\n- 场景配置管理\n\n认证方式：使用Bearer Token认证，在Header中添加 `Authorization: Bearer {token}`",
    "version": "1.0.0",
    "contact": {
      "name": "Lumina AI API Support",
      "email": "api-support@lumina.ai"
    }
  },
  "servers": [
    {
      "url": "https://api.lumina.ai/v1",
      "description": "生产环境"
    },
    {
      "url": "https://api-staging.lumina.ai/v1",
      "description": "测试环境"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/auth/send-code": {
      "post": {
        "tags": ["Authentication"],
        "summary": "发送验证码",
        "description": "向指定手机号发送登录验证码",
        "operationId": "sendVerificationCode",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendCodeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "验证码发送成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendCodeResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "description": "请求过于频繁，请稍后再试",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "手机号登录",
        "description": "使用手机号和验证码登录",
        "operationId": "login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/auth/wechat-login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "微信登录",
        "description": "使用微信授权码登录",
        "operationId": "wechatLogin",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeChatLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "summary": "登出",
        "description": "退出登录，清除token",
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "登出成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/ai/quiz": {
      "post": {
        "tags": ["AI"],
        "summary": "提交AI问答答案",
        "description": "提交用户完成的AI问答答案，用于生成个性化推荐",
        "operationId": "submitQuizAnswers",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizSubmissionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "提交成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizSubmissionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/ai/recommendations": {
      "get": {
        "tags": ["AI"],
        "summary": "获取AI推荐",
        "description": "基于用户问答结果获取个性化推荐场景和功能",
        "operationId": "getRecommendations",
        "parameters": [
          {
            "name": "quizSessionId",
            "in": "query",
            "description": "问答会话ID（如果提供则基于问答结果，否则基于用户历史）",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/images/upload": {
      "post": {
        "tags": ["Images"],
        "summary": "上传图片",
        "description": "上传单张或多张图片，返回图片ID和URL",
        "operationId": "uploadImages",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["files"],
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "图片文件（支持JPG, PNG, WebP，单次最多100张）"
                  },
                  "sceneType": {
                    "type": "string",
                    "enum": ["taobao", "douyin", "xiaohongshu", "amazon", "custom"],
                    "description": "场景类型（可选）"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "上传成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "413": {
            "description": "文件过大",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/images/process": {
      "post": {
        "tags": ["Images"],
        "summary": "处理单张图片",
        "description": "对单张图片进行AI处理（抠图、背景替换、光效等）",
        "operationId": "processImage",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessImageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "处理任务已创建",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessTaskResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/images/batch-process": {
      "post": {
        "tags": ["Images"],
        "summary": "批量处理图片",
        "description": "批量处理多张图片，使用相同的处理参数",
        "operationId": "batchProcessImages",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchProcessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "批量处理任务已创建",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchProcessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/images/process/{taskId}/status": {
      "get": {
        "tags": ["Images"],
        "summary": "获取处理进度",
        "description": "查询图片处理任务的状态和进度",
        "operationId": "getProcessStatus",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "description": "处理任务ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessStatusResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/images/process/{taskId}/result": {
      "get": {
        "tags": ["Images"],
        "summary": "获取处理结果",
        "description": "获取处理完成后的结果图片URL",
        "operationId": "getProcessResult",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "description": "处理任务ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessResultResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/images/{imageId}/download": {
      "get": {
        "tags": ["Images"],
        "summary": "下载图片",
        "description": "下载处理后的图片（支持不同质量等级）",
        "operationId": "downloadImage",
        "parameters": [
          {
            "name": "imageId",
            "in": "path",
            "required": true,
            "description": "图片ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quality",
            "in": "query",
            "description": "图片质量（60-100，默认85）",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 60,
              "maximum": 100,
              "default": 85
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "图片格式（jpg, png, webp，默认jpg）",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["jpg", "png", "webp"],
              "default": "jpg"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "下载成功",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "使用JWT Token进行认证，格式：Bearer {token}"
      }
    },
    "schemas": {
      "SendCodeRequest": {
        "type": "object",
        "required": ["phoneNumber"],
        "properties": {
          "phoneNumber": {
            "type": "string",
            "pattern": "^1[3-9]\\d{9}$",
            "description": "手机号码（11位，以1开头）",
            "example": "13812345678"
          }
        }
      },
      "SendCodeResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "验证码已发送"
          },
          "expiresIn": {
            "type": "integer",
            "description": "验证码有效期（秒）",
            "example": 300
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["phoneNumber", "verificationCode"],
        "properties": {
          "phoneNumber": {
            "type": "string",
            "pattern": "^1[3-9]\\d{9}$",
            "description": "手机号码",
            "example": "13812345678"
          },
          "verificationCode": {
            "type": "string",
            "pattern": "^\\d{6}$",
            "description": "6位验证码",
            "example": "123456"
          }
        }
      },
      "WeChatLoginRequest": {
        "type": "object",
        "required": ["code"],
        "properties": {
          "code": {
            "type": "string",
            "description": "微信授权码",
            "example": "081abc123def456"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT认证令牌",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "description": "刷新令牌",
            "example": "refresh_token_here"
          },
          "user": {
            "$ref": "#/components/schemas/UserProfile"
          },
          "isNewUser": {
            "type": "boolean",
            "description": "是否为新用户",
            "example": true
          },
          "expiresIn": {
            "type": "integer",
            "description": "Token过期时间（秒）",
            "example": 7200
          }
        }
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "user_123456"
          },
          "phoneNumber": {
            "type": "string",
            "example": "138****5678"
          },
          "nickname": {
            "type": "string",
            "example": "创作者"
          },
          "avatar": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "example": "https://cdn.lumina.ai/avatars/user_123456.jpg"
          },
          "isPro": {
            "type": "boolean",
            "description": "是否为专业版会员",
            "example": true
          },
          "membershipType": {
            "type": "string",
            "enum": ["free", "monthly", "annual", "enterprise"],
            "example": "annual"
          },
          "membershipExpiry": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2025-01-28T00:00:00Z"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T00:00:00Z"
          }
        }
      },
...
      "QuizSubmissionRequest": {
        "type": "object",
        "required": ["answers"],
        "properties": {
          "answers": {
            "type": "object",
            "description": "问答答案，key为问题序号（1-based），value为选项ID",
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "1": "clothing",
              "2": "taobao",
              "3": "minimal"
            }
          }
        }
      },
      "QuizSubmissionResponse": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "问答会话ID",
            "example": "quiz_session_abc123"
          },
          "message": {
            "type": "string",
            "example": "提交成功"
          }
        }
      },
      "RecommendationsResponse": {
        "type": "object",
        "properties": {
          "primaryRecommendation": {
            "$ref": "#/components/schemas/Recommendation"
          },
          "alternatives": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            }
          },
          "recommendedFeatures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Feature"
            }
          }
        }
      },
      "Recommendation": {
        "type": "object",
        "properties": {
          "sceneType": {
            "type": "string",
            "enum": ["taobao", "douyin", "xiaohongshu", "amazon", "custom"],
            "example": "taobao"
          },
          "sceneName": {
            "type": "string",
            "example": "电商主图"
          },
          "matchPercentage": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "example": 98
          },
          "previewImage": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "example": "https://cdn.lumina.ai/previews/taobao.jpg"
          },
          "description": {
            "type": "string",
            "example": "基于您的选择，这个场景最适合您的需求"
          }
        }
      },
      "Feature": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "smart_cutout"
          },
          "name": {
            "type": "string",
            "example": "智能抠图"
          },
          "description": {
            "type": "string",
            "example": "AI精准识别主体，一键移除背景"
          },
          "icon": {
            "type": "string",
            "example": "sparkles"
          }
        }
      },
      "UploadResponse": {
        "type": "object",
        "properties": {
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UploadedImage"
            }
          }
        }
      },
...
      "ProcessImageRequest": {
        "type": "object",
        "required": ["imageId", "operations"],
        "properties": {
          "imageId": {
            "type": "string",
            "description": "要处理的图片ID",
            "example": "img_abc123"
          },
          "operations": {
            "type": "array",
            "description": "处理操作列表",
            "items": {
              "$ref": "#/components/schemas/ImageOperation"
            }
          },
          "outputSize": {
            "type": "string",
            "pattern": "^\\d+x\\d+$",
            "description": "输出尺寸（如：2000x2000）",
            "example": "2000x2000"
          },
          "quality": {
            "type": "integer",
            "minimum": 60,
            "maximum": 100,
            "default": 85,
            "description": "输出质量（60-100）",
            "example": 85
          },
          "edgeSmoothing": {
            "type": "boolean",
            "default": true,
            "description": "是否启用边缘平滑",
            "example": true
          },
          "sceneType": {
            "type": "string",
            "enum": ["taobao", "douyin", "xiaohongshu", "amazon", "custom"],
            "description": "场景类型",
            "example": "taobao"
          }
        }
      },
...
      "ProcessTaskResponse": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string",
            "description": "处理任务ID",
            "example": "task_xyz789"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "processing", "completed", "failed"],
            "example": "pending"
          },
          "estimatedTime": {
            "type": "integer",
            "description": "预计处理时间（秒）",
            "example": 5
          }
        }
      },
      "BatchProcessRequest": {
        "type": "object",
        "required": ["imageIds", "operations"],
        "properties": {
          "imageIds": {
            "type": "array",
            "description": "要处理的图片ID列表（最多100张）",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "maxItems": 100,
            "example": ["img_abc123", "img_def456"]
          },
          "operations": {
            "type": "array",
            "description": "处理操作列表（应用到所有图片）",
            "items": {
              "$ref": "#/components/schemas/ImageOperation"
            }
          },
          "outputSize": {
            "type": "string",
            "pattern": "^\\d+x\\d+$",
            "example": "2000x2000"
          },
          "quality": {
            "type": "integer",
            "minimum": 60,
            "maximum": 100,
            "default": 85,
            "example": 85
          },
          "edgeSmoothing": {
            "type": "boolean",
            "default": true,
            "example": true
          },
          "sceneType": {
            "type": "string",
            "enum": ["taobao", "douyin", "xiaohongshu", "amazon", "custom"],
            "example": "taobao"
          }
        }
      },
      "BatchProcessResponse": {
        "type": "object",
        "properties": {
          "batchTaskId": {
            "type": "string",
            "description": "批量处理任务ID",
            "example": "batch_task_abc123"
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProcessTaskResponse"
            }
          },
          "totalCount": {
            "type": "integer",
            "example": 2
          }
        }
      },
      "ProcessStatusResponse": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string",
            "example": "task_xyz789"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "processing", "completed", "failed"],
            "example": "processing"
          },
          "progress": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "description": "处理进度（0-100）",
            "example": 65
          },
          "message": {
            "type": "string",
            "nullable": true,
            "description": "状态消息",
            "example": "正在处理中..."
          },
          "estimatedTimeRemaining": {
            "type": "integer",
            "nullable": true,
            "description": "预计剩余时间（秒）",
            "example": 2
          }
        }
      },
      "ProcessResultResponse": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string",
            "example": "task_xyz789"
          },
          "status": {
            "type": "string",
            "enum": ["completed", "failed"],
            "example": "completed"
          },
          "resultImage": {
            "$ref": "#/components/schemas/ProcessedImage"
          },
          "processingTime": {
            "type": "number",
            "description": "处理耗时（秒）",
            "example": 3.2
          },
          "beforeImage": {
            "$ref": "#/components/schemas/UploadedImage"
          },
          "error": {
            "type": "string",
            "nullable": true,
            "description": "错误信息（如果失败）",
            "example": null
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "操作成功"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "INVALID_REQUEST"
              },
              "message": {
                "type": "string",
                "example": "请求参数错误"
              },
              "details": {
                "type": "object",
                "additionalProperties": true,
                "description": "错误详情"
              }
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "未授权，需要登录",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}